+++ crates/store/src/lib.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 mod blocks;
 pub mod db;
 pub mod errors;
@@ -13,3 +15,6 @@ pub const GENESIS_STATE_FILENAME: &str = "genesis.dat";
 
 /// Number of sql statements that each connection will cache.
 const SQL_STATEMENT_CACHE_CAPACITY: usize = 32;
+
+/// How often to run the database maintenance routine.
+const DATABASE_MAINTENANCE_INTERVAL: Duration = Duration::from_secs(24 * 60 * 60);